Fixes and closes #307
