+++ feedback/utils.py
@@ -1,7 +1,12 @@
 # -*- coding: utf-8 -*-
 '''Helper utilities and decorators.'''
 
+import pytz
+
 from flask import flash, request, url_for
+from tzlocal import get_localzone
+
+local_tz = get_localzone()
 
 
 def flash_errors(form, category="warning"):
@@ -24,3 +29,8 @@ def thispage():
     # pass for favicon
     except AttributeError:
         pass
+
+
+def utc_to_local(utc_dt):
+    local_dt = utc_dt.replace(tzinfo=pytz.utc).astimezone(local_tz)
+    return local_tz.normalize(local_dt)  # .normalize might be unnecessary
